function that creates a file.

+++ 0x15-file_io/1-create_file.c
@@ -9,7 +9,7 @@ int create_file(const char *filename, char *text_content)
 {
 	int fd;
 	int countcontent = 0;
-	int countwrite;
+	long countwrite;
 
 	if (filename == NULL)
 	{
